continue with clean ups and comments
